split models into separate files for readability
