+++ src/involvement/models/contact_card_block.py
@@ -1,0 +1,15 @@
+from django.utils.translation import ugettext_lazy as _
+from wagtail.snippets.blocks import SnippetChooserBlock
+
+
+class ContactCardBlock(SnippetChooserBlock):
+    def __init__(self, **kwargs):
+        super(ContactCardBlock, self).__init__(
+            'involvement.ContactCard', **kwargs
+        )
+
+    class Meta:
+        label = _('Contact Card')
+        icon = 'user'
+        template = 'involvement/blocks/contact_card.html'
+        group = _('Meta')
